backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
pyjwt==2.10.1
langgraph==0.2.66
langchain-core==0.3.29
openfga-sdk==0.5.0
